Use static type information to restore type tags
When we unpickle IValues, we lose type information for List[T]
and Dict[K, V]. We can restore this information using the static
type information contained in the top-level Module/Class type.

This ensures that even after serialization we can always get the
dynamic type of an ivalue using its type() method.

emitModuleHook now has a check that the tags in saved/load modules
are the same.

gh-metadata: pytorch pytorch 25447 gh/zdevito/102/head